Try-on/tryon/requirements.txt
accelerate==1.13.0
aiofiles==23.2.1
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
av==12.3.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
diffusers==0.37.0
exceptiongroup==1.3.1
fastapi==0.135.1
ffmpy==1.0.0
filelock==3.20.0
fonttools==4.62.1
fsspec==2025.12.0
fvcore==0.1.5.post20221221
gradio==4.41.0
gradio_client==1.3.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.7.2
idna==3.11
ImageIO==2.37.3
importlib_metadata==9.0.0
importlib_resources==6.5.2
iopath==0.1.10
Jinja2==3.1.3
kiwisolver==1.5.0
lazy-loader==0.5
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.8
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
omegaconf==2.3.0
opencv-python==4.10.0.84
orjson==3.11.7
packaging==26.0
pandas==2.3.3
peft==0.18.1
pillow==10.3.0
portalocker==3.2.0
psutil==7.2.2
pycocotools==2.0.11
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-multipart==0.0.22
pytz==2026.1.post1
pywin32==311
PyYAML==6.0.1
regex==2026.2.28
requests==2.32.5
rich==14.3.3
ruff==0.15.7
safetensors==0.7.0
scikit-image==0.24.0
scipy==1.13.1
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
starlette==1.0.0
sympy==1.13.1
tabulate==0.10.0
termcolor==3.3.0
tifffile==2025.5.10
tokenizers==0.22.2
tomlkit==0.12.0
torch==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.66.4
transformers==5.3.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.42.0
websockets==12.0
yacs==0.1.8
zipp==3.23.0
